Ink levels decrease. Add more tests

+++ src/printers/InkCartridge.java
@@ -12,10 +12,9 @@ public class InkCartridge extends Cartridge {
 	public double getLevel(){
 		return level;
 	}
-	
-	//helper testing method
+
 	public void setLevel(double value){
-		if(value <= (double) 100 && value > (double) 0){
+		if(value <= (double) 100 && value > 0){
 			this.level = value;
 		}else{
 			System.out.println("The value is wrong");
